test(mac): move tests for root to subcrate
