+++ .envrc
@@ -1,0 +1,1 @@
+export RUST_BACKTRACE=1
