+++ tests/lib.rs
@@ -1,6 +1,3 @@
 extern crate octavo;
-extern crate openssl;
-extern crate quickcheck;
 
 mod kdf;
-mod mac;
